Renamed GOSettingsAudioOutput to GOSettingsAudio https://github.com/GrandOrgue/grandorgue/issues/948
